+++ packages/effect/src/data/Tuple.ts
@@ -7,27 +7,8 @@ import * as Equivalence from "../data/Equivalence.ts"
 import * as order from "../data/Order.ts"
 import type { Apply, Lambda } from "../data/Struct.ts"
 import { dual } from "../Function.ts"
-import type { TypeLambda } from "../types/HKT.ts"
 import * as Combiner from "./Combiner.ts"
 import * as Reducer from "./Reducer.ts"
-
-/**
- * A type lambda for tuples with two elements, useful for higher-kinded type operations.
- *
- * @example
- * ```ts
- * import type { Tuple } from "effect/data"
- *
- * // Used internally for type-level operations on 2-tuples
- * type Example = Tuple.Tuple2TypeLambda
- * ```
- *
- * @category Type lambdas
- * @since 4.0.0
- */
-export interface Tuple2TypeLambda extends TypeLambda {
-  readonly type: [this["Out1"], this["Target"]]
-}
 
 /**
  * Constructs a new tuple from the provided values.
@@ -150,38 +131,6 @@ export const omit: {
     return self.filter((_, i) => !toDrop.has(i))
   }
 )
-
-/**
- * Return the first element of a tuple.
- *
- * @example
- * ```ts
- * import * as assert from "node:assert"
- * import { getFirst } from "effect/data/Tuple"
- *
- * assert.deepStrictEqual(getFirst(["hello", 42]), "hello")
- * ```
- *
- * @category Getters
- * @since 2.0.0
- */
-export const getFirst = <L, R>(self: readonly [L, R]): L => self[0]
-
-/**
- * Return the second element of a tuple.
- *
- * @example
- * ```ts
- * import * as assert from "node:assert"
- * import { getSecond } from "effect/data/Tuple"
- *
- * assert.deepStrictEqual(getSecond(["hello", 42]), 42)
- * ```
- *
- * @category Getters
- * @since 2.0.0
- */
-export const getSecond = <L, R>(self: readonly [L, R]): R => self[1]
 
 /**
  * Appends an element to the end of a tuple.
@@ -410,103 +359,6 @@ export const mapOmit: {
     return self.map((e, i) => (toOmit.has(i) ? e : lambda(e)))
   }
 )
-
-/**
- * Transforms both elements of a tuple using the given functions.
- *
- * @example
- * ```ts
- * import * as assert from "node:assert"
- * import { mapBoth } from "effect/data/Tuple"
- *
- * assert.deepStrictEqual(
- *   mapBoth(["hello", 42], { onFirst: s => s.toUpperCase(), onSecond: n => n.toString() }),
- *   ["HELLO", "42"]
- * )
- * ```
- *
- * @category Mapping
- * @since 2.0.0
- */
-export const mapBoth: {
-  <L1, L2, R1, R2>(options: {
-    readonly onFirst: (e: L1) => L2
-    readonly onSecond: (a: R1) => R2
-  }): (self: readonly [L1, R1]) => [L2, R2]
-  <L1, R1, L2, R2>(self: readonly [L1, R1], options: {
-    readonly onFirst: (e: L1) => L2
-    readonly onSecond: (a: R1) => R2
-  }): [L2, R2]
-} = dual(
-  2,
-  <L1, R1, L2, R2>(
-    self: readonly [L1, R1],
-    { onFirst, onSecond }: {
-      readonly onFirst: (e: L1) => L2
-      readonly onSecond: (a: R1) => R2
-    }
-  ): [L2, R2] => [onFirst(self[0]), onSecond(self[1])]
-)
-
-/**
- * Transforms the first component of a tuple using a given function.
- *
- * @example
- * ```ts
- * import * as assert from "node:assert"
- * import { mapFirst } from "effect/data/Tuple"
- *
- * assert.deepStrictEqual(
- *   mapFirst(["hello", 42], s => s.toUpperCase()),
- *   ["HELLO", 42]
- * )
- * ```
- *
- * @category Mapping
- * @since 2.0.0
- */
-export const mapFirst: {
-  <L1, L2>(f: (left: L1) => L2): <R>(self: readonly [L1, R]) => [L2, R]
-  <L1, R, L2>(self: readonly [L1, R], f: (left: L1) => L2): [L2, R]
-} = dual(2, <L1, R, L2>(self: readonly [L1, R], f: (left: L1) => L2): [L2, R] => [f(self[0]), self[1]])
-
-/**
- * Transforms the second component of a tuple using a given function.
- *
- * @example
- * ```ts
- * import * as assert from "node:assert"
- * import { mapSecond } from "effect/data/Tuple"
- *
- * assert.deepStrictEqual(
- *   mapSecond(["hello", 42], n => n.toString()),
- *   ["hello", "42"]
- * )
- * ```
- *
- * @category Mapping
- * @since 2.0.0
- */
-export const mapSecond: {
-  <R1, R2>(f: (right: R1) => R2): <L>(self: readonly [L, R1]) => [L, R2]
-  <L, R1, R2>(self: readonly [L, R1], f: (right: R1) => R2): [L, R2]
-} = dual(2, <L, R1, R2>(self: readonly [L, R1], f: (right: R1) => R2): [L, R2] => [self[0], f(self[1])])
-
-/**
- * Swaps the two elements of a tuple.
- *
- * @example
- * ```ts
- * import * as assert from "node:assert"
- * import { flip } from "effect/data/Tuple"
- *
- * assert.deepStrictEqual(flip(["hello", 42]), [42, "hello"])
- * ```
- *
- * @category Tuple2
- * @since 4.0.0
- */
-export const flip = <L, R>(self: readonly [L, R]): [R, L] => [self[1], self[0]]
 
 /**
  * Creates an `Equivalence` for tuples by comparing corresponding elements using the provided `Equivalence`s.
